refactor: 딜러의 이름을 기본값으로 설정하는 기능 수정
- name 속성을 상속 불가능한 타입으로 설정 후 기본값 부여
