+++ src/main/kotlin/blackjack/domain/CardGenerator.kt
@@ -1,6 +1,8 @@
-package blackjack.domain.card
+package blackjack.domain
 
-import blackjack.domain.Generator
+import blackjack.domain.card.Card
+import blackjack.domain.card.CardNumber
+import blackjack.domain.card.CardShape
 
 class CardGenerator(private val randomGenerator: Generator) {
